Add main content styling
The main content div is divided into a grid. Both projects,
announcements and trending, as well as their headers, have their own
cells.

Grid is also used to divide the projects into two columns.
Combinations of grid and flex are used in both projects, announcements
and trending to place the elements correctly.