Change :hover and :active effects
Removes the transform effect from all elements other than the modifier
buttons. Changes background effects for the projects.

Also adds a left border on :hover for sidebar links and the projects.
Because all elements have "box-sizing: border-box", all the content
shifts to the right when the border is applied, which makes the elements
stand out even more.