Add toggle effects to project buttons
Uses JavaScript to detect mousedown event, which adds or removes the
'toggled' class from the button. When toggled, the star button becomes
yellow and the watch later button becomes blue using the filter
function.

Also transforms buttons on :hover and :active to give the user
feedback when using the buttons.