Move project buttons, changes grid structure
Moves the project buttons to prevent a button click from activating the
link on the project. This is done by changing to relative positioning
for the project div and absolute positioning for the project button div.

Also changes grid structure for both the main content grid and the grid
for each project. The main content grid now uses minmax(auto, 350px) to
limit announcements and trending to 350px, so the scroll overflow
actually works. An additional row is also added to allow projects to
expand when needed. The grid for each project has a strict height of
--distance-large (2rem) to give the project buttons the correct space.